♻️ refactor: router 작동 하지 않는 부분 수정
